refactor: add decorator and middleware in create_app
