+++ .env.test.sample
@@ -1,9 +1,9 @@
 # frontend
-GRAPHQL_API_URL=http://localhost:5000/v1/graphql
-BACKEND_API_URL=http://localhost:8000
-BACKEND_API_ACTION_URL=http://host.docker.internal:8000
+GRAPHQL_API_URL=http://localhost:5001/v1/graphql
+BACKEND_API_URL=http://localhost:8001
+BACKEND_API_ACTION_URL=http://host.docker.internal:8001
 PUBLIC_MATOMO_URL=
-PUBLIC_MATOMO_SITE_ID=
+PUBLIC_MATOMO_SITE_ID=SITE_ID
 PUBLIC_CRISP_WEBSITE_ID=
 PUBLIC_SENTRY_DSN=
 PUBLIC_SENTRY_ENVIRONMENT=
